map: support big maps
support block dictionaries bigger than 8bit. This means
that we drop from 25% to 50% compression ration, but we
need to support big maps.

+++ engine/map.c
@@ -28,12 +28,12 @@
  *   width of the map is needed
  *   out buffer needs to be 4 times data_size
  */
-void map_inflate(const uint8_t * dict, const uint8_t * in, uint8_t * out, uint16_t data_size,
+void map_inflate(const uint8_t * dict, const uint16_t * in, uint8_t * out, uint16_t data_size,
 		 uint8_t w)
 {
 	uint8_t col = 0;
 	unsigned int idx;
-	uint8_t *src;
+	uint16_t *src;
 	uint8_t *dst = out;
 	/* FIXME: optimize this */
 	for (src = in; src < in + data_size; src++) {
@@ -55,11 +55,11 @@ void map_inflate(const uint8_t * dict, const uint8_t * in, uint8_t * out, uint16
  * Inflate only a 32x24 window of the map to a buffer, useful if the map is too big
  * to be decompressed entirely.
  */
-void __map_inflate_screen(const uint8_t * dict, const uint8_t * in, uint8_t * out, uint8_t w, uint8_t vpx, uint8_t vpy)
+void __map_inflate_screen(const uint8_t * dict, const uint16_t * in, uint8_t * out, uint8_t w, uint8_t vpx, uint8_t vpy)
 {
 	unsigned int idx;
-	uint8_t *base = in + vpx / 2 + (vpy / 2) * w / 2;
-	uint8_t *src = base;
+	uint16_t *base = in + vpx / 2 + (vpy / 2) * w / 2;
+	uint16_t *src = base;
 	uint8_t *dst = out;
 	int x,y;
